rename: LoginTypeScreen -> LoginScreen
